use broadcast_and_wait; add FLComponentHelper class
