+++ examples/advanced/cifar10/cifar10-sim/jobs/cifar10_fedavg/cifar10_fedavg/config/config_fed_server.json
@@ -2,8 +2,8 @@
     "format_version": 2,
     "min_clients": 8,
     "num_rounds": 50,
-    "TRAIN_SPLIT_ROOT": "/tmp/cifar10_splits/cifar10_fedavg_alpha0.1_731413aa-94c6-4d12-b643-247f673bb9c7",
-    "alpha": 0.1,
+    "TRAIN_SPLIT_ROOT": "/tmp/cifar10_splits/cifar10_fedavg_alpha1.0_edc17139-c263-4609-bb0f-5972747f3af5",
+    "alpha": 1.0,
     "server": {
         "heart_beat_timeout": 600
     },
