+++ examples/advanced/cifar10/cifar10-sim/jobs/cifar10_fedavg/meta.json
@@ -1,5 +1,5 @@
 {
-    "name": "cifar10_fedavg_alpha0.1",
+    "name": "cifar10_fedavg_alpha1.0",
     "deploy_map": {
         "cifar10_fedavg": [
             "@ALL"
